requirements.txt
about-time==4.2.1
alive-progress==3.1.5
grapheme==0.6.0
py-enigma==0.1
tqdm==4.66.1
